% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{launch}
\alias{launch}
\title{Launch MCNV2 Shiny app}
\usage{
launch(bedtools_path = NULL, results_dir = NULL, python_env = "r-MCNV2")
}
\arguments{
\item{bedtools_path}{Full path to bedtools executable}

\item{results_dir}{Directory to save results}

\item{python_env}{Name of the Python virtualenv (default 'r-MCNV2')}
}
\description{
Launch MCNV2 Shiny app
}
